requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.1
seaborn==0.11.2
scikit-learn==1.0.2
joblib==1.1.1
requests==2.28.1
streamlit==1.16.0
streamlit-lottie==0.0.3
Pillow==9.3.0
altair==4.2.2
